requirements.txt
aiofiles==0.3.1
httptools==0.0.9
PyYAML==3.12
sanic==0.5.4
ujson==1.35
uvloop==0.8.0
websockets==3.3
